+++ src/target.rs
@@ -61,6 +61,7 @@ pub enum Arch {
     S390X,
     Wasm32,
     Riscv64,
+    Mips64el,
 }
 
 impl fmt::Display for Arch {
@@ -76,6 +77,7 @@ impl fmt::Display for Arch {
             Arch::S390X => write!(f, "s390x"),
             Arch::Wasm32 => write!(f, "wasm32"),
             Arch::Riscv64 => write!(f, "riscv64"),
+            Arch::Mips64el => write!(f, "mips64el"),
         }
     }
 }
@@ -93,6 +95,7 @@ fn get_supported_architectures(os: &Os) -> Vec<Arch> {
             Arch::X86,
             Arch::X86_64,
             Arch::Riscv64,
+            Arch::Mips64el,
         ],
         Os::Windows => vec![Arch::X86, Arch::X86_64, Arch::Aarch64],
         Os::Macos => vec![Arch::Aarch64, Arch::X86_64],
@@ -131,7 +134,7 @@ impl Target {
     ///
     /// Fails if the target triple isn't supported
     pub fn from_target_triple(target_triple: Option<String>) -> Result<Self> {
-        use target_lexicon::{Architecture, ArmArchitecture, OperatingSystem};
+        use target_lexicon::{Architecture, ArmArchitecture, Mips64Architecture, OperatingSystem};
 
         let host_triple = get_host_target()?;
         let (platform, triple) = if let Some(ref target_triple) = target_triple {
@@ -173,6 +176,7 @@ impl Target {
             Architecture::S390x => Arch::S390X,
             Architecture::Wasm32 => Arch::Wasm32,
             Architecture::Riscv64(_) => Arch::Riscv64,
+            Architecture::Mips64(Mips64Architecture::Mips64el) => Arch::Mips64el,
             unsupported => bail!("The architecture {} is not supported", unsupported),
         };
 
@@ -362,6 +366,7 @@ impl Target {
             Arch::S390X => "s390x",
             Arch::Wasm32 => "wasm32",
             Arch::Riscv64 => "riscv64",
+            Arch::Mips64el => "mips64",
         }
     }
 
@@ -388,7 +393,7 @@ impl Target {
                 PlatformTag::manylinux2014()
             }
             Arch::X86 | Arch::X86_64 => PlatformTag::manylinux2010(),
-            Arch::Armv6L | Arch::Wasm32 | Arch::Riscv64 => PlatformTag::Linux,
+            Arch::Armv6L | Arch::Wasm32 | Arch::Riscv64 | Arch::Mips64el => PlatformTag::Linux,
         }
     }
 
@@ -400,7 +405,8 @@ impl Target {
             | Arch::Powerpc64Le
             | Arch::X86_64
             | Arch::S390X
-            | Arch::Riscv64 => 64,
+            | Arch::Riscv64
+            | Arch::Mips64el => 64,
             Arch::Armv6L | Arch::Armv7L | Arch::X86 | Arch::Wasm32 => 32,
         }
     }